Changed email_body.py to api_functions.py
Result from people API was better than csv file which
resulted in changing the initial plan involving csv files.
Therefore, information.py is obsolete and must be updated to
suit the changes.

changes made in api_functions.py:
authenticate() is created in such a way that it can be applied
for both gmail and people(contacts) by changing the third argument.
Check documentation string for more information on the function.

get_contacts() does what it's name suggests. This function gets
the contacts by sending a request to People API and saves the
contacts in contacts.json file. Check doc string for more information
on the function.